backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
redis==5.0.1
openai==1.3.5
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-socketio==5.10.0
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.25.1
PyPDF2==3.0.1
docx2txt==0.8
websockets==12.0

# Enhanced AI Features
llama-index==0.8.54
langchain==0.0.335
transformers==4.34.0
torch==2.1.0
sentence-transformers==2.2.2

# Blockchain Integration
web3==6.11.1
eth-account==0.9.0
eth-typing==3.5.0
eth-utils==2.3.0

# IPFS Storage
ipfshttpclient==0.8.0
multiaddr==0.0.9

# OCR & Document Processing
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python==4.8.1.78

# Machine Learning
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.2

# Speech Recognition
SpeechRecognition==3.10.0
pyttsx3==2.90

# Security
cryptography==41.0.4
pyOpenSSL==23.2.0
eth-hash==0.5.2
eth-keyfile==0.6.1

# Database
sqlalchemy==2.0.22
asyncpg==0.28.0
alembic==1.12.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
